removed sub directories
