+++ integrations/statuspage/webhook/registry.py
@@ -1,15 +1,13 @@
 from port_ocean.context.ocean import ocean
 
 from webhook.consts import WEBHOOK_PATH
-from webhook.webhook_processors.incident_events.incident_update_webhook_processor import (
+from webhook.webhook_processors.incident_update_webhook_processor import (
     IncidentUpdateWebhookProcessor,
 )
-from webhook.webhook_processors.incident_events.incident_webhook_processor import (
+from webhook.webhook_processors.incident_webhook_processor import (
     IncidentWebhookProcessor,
 )
-from webhook.webhook_processors.page_events.page_webhook_processor import (
-    PageWebhookProcessor,
-)
+from webhook.webhook_processors.page_webhook_processor import PageWebhookProcessor
 
 
 def register_live_events_webhooks() -> None:
