+++ integrations/statuspage/webhook/webhook_processors/incident_update_webhook_processor.py
@@ -8,20 +8,26 @@ from port_ocean.core.handlers.webhook.webhook_event import (
     WebhookEventRawResults,
 )
 from webhook.consts import WebhookPayloadKey
-from webhook.webhook_processors.incident_events.base_processor import (
+from webhook.webhook_processors.incident_base_processor import (
     BaseIncidentWebhookProcessor,
 )
 
 
-class IncidentWebhookProcessor(BaseIncidentWebhookProcessor):
+class IncidentUpdateWebhookProcessor(BaseIncidentWebhookProcessor):
     async def should_process_event(self, event: WebhookEvent) -> bool:
-        return WebhookPayloadKey.INCIDENT in event.payload
+        incident = self.get_incident(event.payload)
+        return incident is not None and WebhookPayloadKey.INCIDENT_UPDATES in incident
 
     async def get_matching_kinds(self, event: WebhookEvent) -> list[str]:
-        return [ObjectKind.INCIDENT]
+        return [ObjectKind.INCIDENT_UPDATE]
 
     async def validate_payload(self, payload: EventPayload) -> bool:
-        return self.get_incident(payload) is not None
+        incident = self.get_incident(payload)
+        if incident is None:
+            return False
+
+        incident_updates = incident.get(WebhookPayloadKey.INCIDENT_UPDATES)
+        return isinstance(incident_updates, list) and len(incident_updates) > 0
 
     async def handle_event(
         self, payload: EventPayload, resource_config: ResourceConfig
@@ -32,11 +38,12 @@ class IncidentWebhookProcessor(BaseIncidentWebhookProcessor):
                 updated_raw_results=[], deleted_raw_results=[]
             )
 
+        incident_updates = incident[WebhookPayloadKey.INCIDENT_UPDATES]
         logger.info(
-            f"Processing Statuspage incident webhook for incident: {incident.get('id')}"
+            f"Processing Statuspage incident update webhook with {len(incident_updates)} updates"
         )
 
         return WebhookEventRawResults(
-            updated_raw_results=[incident],
+            updated_raw_results=incident_updates,
             deleted_raw_results=[],
         )
